Move API key actor mapping into auth layer
