+++ api/app/src/domain/actor.ts
@@ -1,4 +1,3 @@
-import type { ApiKeyAuthResult } from "../auth/api-key";
 import type { AuthIdentity, OrgGate } from "../auth/identity";
 import { AuthzError } from "./errors";
 
@@ -83,16 +82,5 @@ export function actorFromAuthIdentity(
     orgId: identity.orgId,
     source,
     userId: identity.userId,
-  };
-}
-
-export function actorFromApiKeyAuth(auth: ApiKeyAuthResult): Actor {
-  return {
-    createdByUserId: auth.identity.userId,
-    keyId: auth.apiKeyId,
-    kind: "apiKey",
-    orgGate: auth.identity.orgGate,
-    orgId: auth.identity.orgId,
-    scopes: auth.scopes,
   };
 }
